+++ AEPContentAnalytics/Sources/Constants/Core/ContentAnalyticsConstants.swift
@@ -47,7 +47,7 @@ enum ContentAnalyticsConstants {
     static let MAX_WAIT_TIME_MULTIPLIER: Double = 2.5
 
     /// Extension version
-    static let EXTENSION_VERSION = "5.0.2"
+    static let EXTENSION_VERSION = "5.0.3"
 
     /// Extension friendly name
     static let FRIENDLY_NAME = "Content Analytics"